Fix crash when root directory does not exist
We would crash trying to get the prefix for the files if our root
directory did not exist (I found this out by accidentally passing the
wrong path for a nightly build). This fixes that problem by deferring if
we don't find any files.

We should probably also just put the source path into the cloner, too,
since that's the only thing that uses it.