Make uploader use all json
Previously, we were serializing sub-objects (and sub-sub-objects) into
a string, before shoving them into the json object that gets passed
around the message queue. There's no particular need to do that, so
it's kind of dumb extra work we're doing in a couple places. This fixes
that by just passing around dicts serialized as json. Much cleaner, and
easier to read the contents while debugging, too!