Change archive directory name
Normally we'd have been creating timestamped directories for any
netconfig/os pair that finished after the first netconfig/os pair to
finish for a particular srid. Now, the directories are named
srid_netconfig_os, so only re-runs of the whole 3-tuple will result in a
timestamped directory. This is what we want, really.