Make points for a run line up on the graphs
We do this by setting the timestamp for the run (which is used for
graphing purposes) once we know we can actually do the run. We also
sleep for a second in order to avoid having 2 different runs show up
with the same timestamp.

This also has the happy effect of making it so sum(totals) in the sql
queries for the graphs don't end up showing weird spikes on a day when
multiple runs are made. Hooray!