No CLI argument for using browser chrome
The code is being kept in for now, since I don't know which way we're
going to go, but we don't need an option to determine whether or not we
display the browser chrome, as it's going to be either one or the other.
Most likely not, but we have to double check to make sure.