Start work on doing the pcap from the servers
This is in place of doing it on the clients. Linux seems to be the only
stable OS for doing pcaps on programmatically, and all the servers are
linux, so let's do all our pcaps on linux!

This contains the server component that will actually be responsible for
running the pcaps, as well as a slight refactor to share some code
between that and the web-based MQ that windows uses.