Add dns check process
This lets us make sure we won't run tests if the DNS failed to update.
This will be much better than having giant spikes in the graphs when
things go wonky. Also, this will email me, so I won't have to check the
graphs every day to see if something broke on the DNS side of things.