Add OS X user startup for srworker
The srworker launch plist (which needs to be moved from system-wide to
user-specific location) has been modified to open a .command file. This
will cause a terminal to be spawned when the user is logged in, running
the contents of that .command file as a shell script. In this case, it
sets the python path and launches srworker inside the terminal.