Make something that will actually work for srworker on osx
Instead of running some login agent to open a command file (or many
other things I tried), we have to create a terminal profile that will
run stone ridge, set that profile as the default, and add the terminal
as a startup item. The shell script added in this commit contains the
commands to do that.